types moved, toSorted() used
