+++ src/constants/filtersCategory.ts
@@ -1,0 +1,5 @@
+export enum Category {
+    ELECTRONICS = 'Electronics',
+    SHOES = 'Shoes',
+    CLOTHES = 'Clothes',
+}
