+++ src/constants/filtersSortOption.ts
@@ -1,0 +1,6 @@
+export enum SortOption {
+    PRICE_HIGH_TO_LOW = 'priceDesc',
+    PRICE_LOW_TO_HIGH = 'priceAsc',
+    NEWEST = 'dateNewest',
+    OLDEST = 'dateOldest',
+}
