+++ tsconfig.json
@@ -2,7 +2,7 @@
   "compilerOptions": {
     "target": "ES2020",
     "useDefineForClassFields": true,
-    "lib": ["es2020", "dom", "dom.iterable"],
+    "lib": ["ESNext.Array", "es2020", "dom", "dom.iterable"],
     "types": ["node"],
     "allowJs": true,
     "module": "ESNext",
